BUILDER: Updated to switch repos & branches; Copy completed files
Added ability to specify a git account, repo, and branch for the
3dr-arm-yocto-bsp repo.  This will allow you to run the build from the
default OpenSolo repo or any fork. It will also let you specify the
branch to look in.  Once complete, the two tar.gz files are copied into
a date stamped folder and an MD5sum created on them. This will save some
steps when getting them put on Solo. And it allows for file integrity
checking.